Added actualWeather method and reduced unnecessary fields

+++ src/main/java/com/backend/holydaystravel/openweather/domain/Clouds.java
@@ -8,5 +8,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Clouds {
-    private long all;
+    private int all;
 }
